hungry_calculator_backend/backend/requirements.txt
psycopg2==2.9.9
django==4.1.1
djangorestframework==3.12.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
pytz==2024.1
pytest==8.1.2
pytest-django
factory-boy==3.2.1
Faker==15.3.4
